requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-dotenv==1.1.1
ultralytics>=8.3.0
